refactor: restructure Order components and update imports for improved organization
