+++ src/Components/Orders/Order.js
@@ -1,11 +1,13 @@
 import Ordertable from "./Ordertable";
-import { OrderDetails } from "./Orderdetails";
+// import { OrderDetails } from "./Orderdetails";
 
  const Order = () => {
   return (
     <div className="container mx-auto px-4 py-6">
       <h1 className="text-2xl font-bold mb-4">Order Management</h1>
-      <Ordertable data={OrderDetails}/>
+      <div>
+      <Ordertable/>
+      </div>
       {/* Other components or content can be added here */}
     </div>
   );
